adds migration script for phrase fk (#41)
updates word entities
