+++ src/main/java/com/savvato/tribeapp/entities/Noun.java
@@ -1,11 +1,32 @@
 package com.savvato.tribeapp.entities;
 
-import javax.persistence.Entity;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name="noun")
-public class Noun extends AbstractWordEntity {
+public class Noun {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    private String word;
+
+    public String getWord() {
+        return word;
+    }
+
+    public void setWord(String word) {
+        this.word = word;
+    }
 
     public Noun(String word) {
         this.setWord(word);
