+++ src/main/resources/db/migration/changelog-202303281903.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
+
+    <changeSet author="audra" id="202303281903-01">
+
+        <addForeignKeyConstraint  baseColumnNames="adverbId"
+                                  baseTableCatalogName="tribeapp_db"
+                                  baseTableName="phrase"
+                                  constraintName="phrase_fk1"
+                                  referencedColumnNames="id"
+                                  referencedTableCatalogName="tribeapp_db"
+                                  referencedTableName="adverb"
+        />
+
+        <addForeignKeyConstraint  baseColumnNames="verbId"
+                                  baseTableCatalogName="tribeapp_db"
+                                  baseTableName="phrase"
+                                  constraintName="phrase_fk2"
+                                  referencedColumnNames="id"
+                                  referencedTableCatalogName="tribeapp_db"
+                                  referencedTableName="verb"
+        />
+
+        <addForeignKeyConstraint  baseColumnNames="prepositionId"
+                                  baseTableCatalogName="tribeapp_db"
+                                  baseTableName="phrase"
+                                  constraintName="phrase_fk3"
+                                  referencedColumnNames="id"
+                                  referencedTableCatalogName="tribeapp_db"
+                                  referencedTableName="preposition"
+        />
+
+        <addForeignKeyConstraint  baseColumnNames="nounId"
+                                  baseTableCatalogName="tribeapp_db"
+                                  baseTableName="phrase"
+                                  constraintName="phrase_fk4"
+                                  referencedColumnNames="id"
+                                  referencedTableCatalogName="tribeapp_db"
+                                  referencedTableName="noun"
+        />
+
+    </changeSet>
+
+
+</databaseChangeLog>
+
