+++ src/main/resources/db/migration/changelog-master.xml
@@ -14,6 +14,7 @@
 	<include file="changelog-202303161319.xml" 	relativeToChangelogFile="true"/>
 	<include file="changelog-202202221108.xml" 	relativeToChangelogFile="true"/>
 	<include file="changelog-202302231158.xml" 	relativeToChangelogFile="true"/>
+	<include file="changelog-202303281903.xml" 	relativeToChangelogFile="true"/>
 
 </databaseChangeLog>
